feat: rebrand a Shoes Store, migra DB a shoesstore y arregla integracion full-stack
Backend:
- IDs de entidades Integer -> Long (alineado con JpaRepository<,Long>)
- precio Double -> BigDecimal (corrige crash Hibernate "scale has no meaning")
- relaciones @ManyToOne LAZY -> EAGER (serializacion JSON anidada)
- quita @CrossOrigin de ProductoController (CORS solo en WebConfig)
- application.properties: password via DB_PASSWORD (sin secreto en repo)
- base de datos zapateria -> shoesstore

Frontend:
- rebrand a SHOES.STORE: logo con punto acento, fuente Poppins
- tokens de diseno oficiales del sitio Shoes.Store (monocromo #0d0d0d)
- textos de sneakers/tenis, footer y titulo Shoes Store

Docs:
- CLAUDE.md refleja estado resuelto, setup de entorno (JDK 17) y nombres
- .gitignore ignora nbproject del frontend

+++ zapateria-backend/src/main/java/com/zapateria/model/Categoria.java
@@ -14,7 +14,7 @@ public class Categoria {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID_CATEGORIA")
-    private Integer idCategoria;
+    private Long idCategoria;
 
     /** Nombre de la categoría (obligatorio, máximo 100 caracteres). */
     @Column(name = "NOMBRE", nullable = false, length = 100)
@@ -24,15 +24,15 @@ public class Categoria {
     public Categoria() {}
 
     /** Constructor con todos los campos */
-    public Categoria(Integer idCategoria, String nombre) {
+    public Categoria(Long idCategoria, String nombre) {
         this.idCategoria = idCategoria;
         this.nombre = nombre;
     }
 
     // ========== Getters y Setters ==========
 
-    public Integer getIdCategoria() { return idCategoria; }
-    public void setIdCategoria(Integer idCategoria) { this.idCategoria = idCategoria; }
+    public Long getIdCategoria() { return idCategoria; }
+    public void setIdCategoria(Long idCategoria) { this.idCategoria = idCategoria; }
 
     public String getNombre() { return nombre; }
     public void setNombre(String nombre) { this.nombre = nombre; }
